finish home

+++ techstore/src/main/java/com/techstore/dto/response/DiscountResponse.java
@@ -1,6 +1,8 @@
 package com.techstore.dto.response;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -9,8 +11,8 @@ public class DiscountResponse {
     private String code;
     private String description;
     private Integer discountPercent;
-    private Double minOrderAmount;
-    private Double maxDiscountAmount;
+    private BigDecimal minOrderAmount;
+    private BigDecimal maxDiscountAmount;
     private Integer quantity;
     private LocalDate startDate;
     private LocalDate endDate;
